[FIX] product: remove digits precision from uom rounding field
Removed the 'digits_compute' precision on rounding,
to use the complete natural precision of NUMERIC types,
preserving all significant digits.

Modified the view to display the product UOM rounding with a 5 digits value by default.
This value is for usability purpose only, the field still accepts bigger precision, by
setting the `digits` option on the field in the form view.