payment(_*): show incompatible payment methods/providers in debug
Since 17.0 payment methods were introduced to the user and it became
more difficult to debug them in the payment form. For that reason with
this commit we added the availability report in payment form that can
be displayed in debug mode. The report includes information about
compatibility of each payment method and each provider for the given
payment context and the reason if it is incompatible.

task-2882668

closes odoo/odoo#151040

Related: odoo/enterprise#55409
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>