payment_paypal: add idempotency key for paypal requests
An idempotency key was added to avoid processing a payment request twice.
If the payment attempt fails and result wasn't received,
sending the same request again will return the result of the first executed request.
If User decides to retry, another request is made with the same parameters
but a different idempotent id, thus resulting in a new request.

task-2894752

closes odoo/odoo#189150

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>